backend/requirements.txt
# Core dependencies
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0

# YouTube transcript extraction
youtube-transcript-api==0.6.1
yt-dlp==2023.10.13

# Audio transcription fallback
faster-whisper==0.9.0
torch==2.1.0
torchaudio==2.1.0

# Text embedding and processing
sentence-transformers==2.2.2
transformers==4.35.0
numpy==1.24.3

# Zoom API
pyzoom==0.1.4

# Utilities
webvtt-py==0.4.6
beautifulsoup4==4.12.2
lxml==4.9.3
